PDB.fast and PDB.shared automatically switch between native ints and BigIntegers, depending on the size of the number. Due to a bug in the implementation this may lead to incorrect behaviour in some cases.
Applied patch from Bug #294973 to fix.

+++ src/org/eclipse/imp/pdb/facts/impl/shared/SharedIntegerValue.java
@@ -165,7 +165,7 @@ public class SharedIntegerValue extends IntegerValue implements IShareable{
 	}
 	
 	public IInteger negate(){
-		return SharedValueFactory.getInstance().integer((~value) + 1);
+		return SharedValueFactory.getInstance().integer((~((long) value)) + 1);
 	}
